+++ packages/core/src/runtime/Test.ts
@@ -13,7 +13,8 @@ import InnerObserver from './test-observers/Inner'
 import { AnyErrorData, EmptyErrorData, AssertionErrorData } from './errors/Types'
 import { StructuredError } from '../utils/StructuredError'
 
-import { Step, ConditionFn } from './Step'
+import { Step, ConditionFn, StepRecoveryObject, RecoverWith } from './Step'
+import { Looper } from '../Looper'
 
 import { CancellationToken } from '../utils/CancellationToken'
 
@@ -32,6 +33,7 @@ const debug = require('debug')('element:runtime:test')
 export default class Test implements ITest {
 	public settings: ConcreteTestSettings
 	public steps: Step[]
+	public recoverySteps: StepRecoveryObject
 
 	public runningBrowser: Browser<Step> | null
 
@@ -44,6 +46,10 @@ export default class Test implements ITest {
 	public iteration = 0
 
 	public failed: boolean
+
+	public stepCount: number
+
+	public recoveryCount: number
 
 	get skipping(): boolean {
 		return this.failed
@@ -59,9 +65,10 @@ export default class Test implements ITest {
 		this.script = script
 
 		try {
-			const { settings, steps } = script
+			const { settings, steps, recoverySteps } = script
 			this.settings = settings as ConcreteTestSettings
 			this.steps = steps
+			this.recoverySteps = recoverySteps
 
 			// Adds output for console in script
 			script.bindTest(this)
@@ -89,12 +96,17 @@ export default class Test implements ITest {
 	 * @return {Promise<void|Error>}
 	 */
 	public async run(iteration?: number): Promise<void> | never {
-		await this.runWithCancellation(iteration || 0, new CancellationToken())
+		await this.runWithCancellation(
+			iteration || 0,
+			new CancellationToken(),
+			new Looper(this.settings),
+		)
 	}
 
 	public async runWithCancellation(
 		iteration: number,
 		cancelToken: CancellationToken,
+		looper: Looper,
 	): Promise<void> {
 		console.assert(this.client, `client is not configured in Test`)
 
@@ -120,6 +132,8 @@ export default class Test implements ITest {
 
 		this.failed = false
 		this.runningBrowser = null
+		this.stepCount = 0
+		this.recoveryCount = 0
 
 		// await this.observer.attachToNetworkRecorder()
 
@@ -164,23 +178,55 @@ export default class Test implements ITest {
 				} catch (err) {
 					console.log(err.message)
 				}
+				if (!condition) this.stepCount += 1
 				return condition
 			}
 
-			debug('running steps')
-			for (const step of this.steps) {
-				const { once, predicate, skip, pending } = step.options
+			const callCondition = (step: Step): boolean => {
+				const { once, skip, pending } = step.options
 				if (pending) {
 					console.log(`(Pending) ${step.name}`)
-					continue
+					this.stepCount += 1
+					return false
 				}
 				if (once && iteration > 1) {
-					continue
+					this.stepCount += 1
+					return false
 				}
 				if (skip) {
 					console.log(`Skip test ${step.name}`)
-					continue
+					this.stepCount += 1
+					return false
 				}
+				return true
+			}
+
+			const callRecovery = async (step: Step): Promise<boolean> => {
+				const { recoveryStep, loopCount } = this.recoverySteps[step.name]
+				const { maxRecovery } = this.settings
+				const settingRecoveryCount = loopCount || maxRecovery || 1
+				if (!recoveryStep || this.recoveryCount >= settingRecoveryCount) return false
+				this.recoveryCount += 1
+				try {
+					const result = await recoveryStep.fn.call(null, browser)
+					if (result === RecoverWith.CONTINUE) return true
+					if (result === RecoverWith.RESTART) {
+						looper.restartLoop()
+						this.stepCount = this.steps.length
+					}
+				} catch (err) {
+					return false
+				}
+				this.failed = false
+				return true
+			}
+
+			debug('running steps')
+			while (this.stepCount < this.steps.length) {
+				const step = this.steps[this.stepCount]
+				const { predicate } = step.options
+				if (!callCondition(step)) continue
+
 				if (predicate) {
 					const condition = await callPredicate(predicate)
 					if (!condition) {
@@ -199,9 +245,12 @@ export default class Test implements ITest {
 				if (cancelToken.isCancellationRequested) return
 
 				if (this.failed) {
+					const result = await callRecovery(step)
+					if (result) continue
 					console.log('failed, bailing out of steps')
 					throw Error('test failed')
 				}
+				this.stepCount += 1
 			}
 		} catch (err) {
 			console.log('error -> failed', err)