+++ packages/element-cli/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.3.0-beta.9](https://github.com/flood-io/element/compare/v1.3.0-beta.8...v1.3.0-beta.9) (2020-06-03)
+
+**Note:** Version bump only for package element-cli
+
+
+
+
+
+# [1.3.0-beta.8](https://github.com/flood-io/element/compare/v1.3.0-beta.7...v1.3.0-beta.8) (2020-06-03)
+
+**Note:** Version bump only for package element-cli
+
+
+
+
+
 # [1.3.0-beta.7](https://github.com/flood-io/element/compare/v1.3.0-beta.6...v1.3.0-beta.7) (2020-06-03)
 
 **Note:** Version bump only for package element-cli